refactor(directory): route 구성이 아닌 컴포넌트는 외부로 분리(가 맞을까?)
