+++ apps/lostark-hands-next/component/news/update-history.tsx
@@ -1,0 +1,24 @@
+import { FC } from 'react';
+
+import { getNoticeApi } from '@/service/news/notice';
+
+import MessagePostList from '@/component/common/message-post-list';
+
+/* @ts-expect-error Async Server Component */
+const UpdateHistory: FC = async () => {
+	const list = await getNoticeApi();
+
+	const updateList = list
+		.filter(({ Title }) => Title.includes('업데이트 내역 안내'))
+		.map(({ Date, Link, Title }) => ({
+			title: Title,
+			type: '업데이트',
+			date: Date,
+			link: Link
+		}))
+		.splice(0, 10);
+
+	return <MessagePostList list={updateList} />;
+};
+
+export default UpdateHistory;
